fts: extract reader/count.rs (unranked match/count kernels)
Phase 2 (cont). Split the unranked match/count kernels — the atom walk
(walk_atoms_match), the phrase-aware match-id/count (atoms_match_ids,
atoms_match_count), and token_match/token_match_count/term_df(s) — out
of the core impl into reader/count.rs as its own impl FtsReader block.
The public count/df entry points stay pub (the superfile wrapper calls
them); walk_atoms_match, build_atom_cursors, fetch_term_postings,
dict_bytes_async, and or_*_unranked are widened to pub(super). Pure
move; FTS + reader + supertable-query suites green.
